add 'sharedProps' and break the component into two
