+++ src/components/FormElements/ContractMethodSelectPicker.tsx
@@ -3,7 +3,7 @@ import { Select } from "@stellar/design-system";
 import { contract } from "@stellar/stellar-sdk";
 
 import { Box } from "@/components/layout/Box";
-import { JsonSchemaForm } from "@/components/JsonSchema/JsonSchemaForm";
+import { JsonSchemaForm } from "@/components/JsonSchema";
 import { ExpandBox } from "@/components/ExpandBox";
 
 import { SorobanInvokeValue } from "@/types/types";
