+++ src/components/JsonSchema/JsonSchemaFormRenderer.tsx
@@ -1,29 +1,18 @@
-import React, { useEffect, useState } from "react";
-import { BASE_FEE, contract } from "@stellar/stellar-sdk";
-import { Button, Card, Icon, Input, Text } from "@stellar/design-system";
+import React, { useState } from "react";
+import { Button, Card, Icon, Input } from "@stellar/design-system";
 import type { JSONSchema7 } from "json-schema";
 import { parse } from "lossless-json";
 
-import { TransactionBuildParams } from "@/store/createStore";
 import { useStore } from "@/store/useStore";
 import { validate } from "@/validate";
 
-import { DereferencedSchemaType } from "@/constants/jsonSchema";
-
-import { dereferenceSchema } from "@/helpers/dereferenceSchema";
-import { getScValsFromSpec } from "@/helpers/getScValsFromSpec";
-import { buildTxWithSorobanData } from "@/helpers/sorobanUtils";
-import { getNetworkHeaders } from "@/helpers/getNetworkHeaders";
 import { arrayItem } from "@/helpers/arrayItem";
-
-import { useRpcPrepareTx } from "@/query/useRpcPrepareTx";
 
 import { Box } from "@/components/layout/Box";
 import { PositiveIntPicker } from "@/components/FormElements/PositiveIntPicker";
 import { LabelHeading } from "@/components/LabelHeading";
-import { ErrorText } from "@/components/ErrorText";
 
-import { AnyObject, SorobanInvokeValue, TxnOperation } from "@/types/types";
+import { AnyObject, SorobanInvokeValue } from "@/types/types";
 
 export const JsonSchemaFormRenderer = ({
   name,
@@ -53,14 +42,21 @@ export const JsonSchemaFormRenderer = ({
     sorobanOperation.params.invoke_contract,
   ) as AnyObject;
 
+  const schemaType = getDefType(schema);
+  const label = path.length > 0 ? getNestedValueLabel(path.join(".")) : name;
+  const labelWithSchemaType = `${label} (${schemaType})`;
+
   const invokeContractBaseProps = {
     contract_id: parsedSorobanOperation.contract_id,
     function_name: parsedSorobanOperation.function_name,
   };
-
-  const schemaType = getDefType(schema);
-  const label = path.length > 0 ? getNestedValueLabel(path.join(".")) : name;
-  const labelWithSchemaType = `${label} (${schemaType})`;
+  const sharedProps = {
+    id: path.join("."),
+    key: path.join("."),
+    label: labelWithSchemaType,
+    value: getNestedValue(parsedSorobanOperation.args, path.join(".")),
+    error: path.length > 0 ? formError[path.join(".")] : formError[name],
+  };
 
   const handleChange = (
     e: React.ChangeEvent<HTMLInputElement>,
@@ -115,6 +111,16 @@ export const JsonSchemaFormRenderer = ({
       [currentPath]: error ? error : "",
     });
   };
+
+  if (schemaType === null || schemaType === undefined) {
+    return (
+      <Box gap="md">
+        <Box gap="md">
+          <div>No schema found</div>
+        </Box>
+      </Box>
+    );
+  }
 
   if (schemaType === "object") {
     return (
@@ -248,28 +254,20 @@ export const JsonSchemaFormRenderer = ({
     case "Address":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getPublicKeyError);
           }}
           infoText={schema.description || ""}
           leftElement={<Icon.User03 />}
           note={<>{schema.description}</>}
+          fieldSize="md"
         />
       );
     case "U32":
       return (
         <PositiveIntPicker
-          id={path.join(".")}
-          key={path.join(".")}
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getU32Error);
           }}
@@ -278,11 +276,7 @@ export const JsonSchemaFormRenderer = ({
     case "U64":
       return (
         <PositiveIntPicker
-          id={path.join(".")}
-          key={path.join(".")}
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getU64Error);
           }}
@@ -291,11 +285,7 @@ export const JsonSchemaFormRenderer = ({
     case "U128":
       return (
         <PositiveIntPicker
-          id={path.join(".")}
-          key={path.join(".")}
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getU128Error);
           }}
@@ -304,11 +294,7 @@ export const JsonSchemaFormRenderer = ({
     case "U256":
       return (
         <PositiveIntPicker
-          id={path.join(".")}
-          key={path.join(".")}
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getU256Error);
           }}
@@ -317,244 +303,69 @@ export const JsonSchemaFormRenderer = ({
     case "I32":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getI32Error);
           }}
+          fieldSize="md"
         />
       );
     case "I64":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getI64Error);
           }}
+          fieldSize="md"
         />
       );
     case "I128":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getI128Error);
           }}
+          fieldSize="md"
         />
       );
     case "I256":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={labelWithSchemaType}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getI256Error);
           }}
+          fieldSize="md"
         />
       );
     case "ScString":
     case "ScSymbol":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={label}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={""}
+          {...sharedProps}
+          error="" // @TODO
           onChange={(e) => {
             // @TODO validate the value via length
             handleChange(e);
           }}
           note={<>{schema.description}</>}
+          fieldSize="md"
         />
       );
     case "DataUrl":
       return (
         <Input
-          id={path.join(".")}
-          key={path.join(".")}
-          fieldSize="md"
-          label={label}
-          value={getNestedValue(parsedSorobanOperation.args, path.join("."))}
-          error={path.length > 0 ? formError[path.join(".")] : formError[name]}
+          {...sharedProps}
           onChange={(e) => {
             handleChange(e, validate.getDataUrlError);
           }}
+          fieldSize="md"
         />
       );
     default:
       return null;
-  }
-};
-
-export const JsonSchemaForm = ({
-  name,
-  value,
-  onChange,
-  spec,
-  funcSchema,
-}: {
-  name: string;
-  value: SorobanInvokeValue;
-  onChange: (value: SorobanInvokeValue) => void;
-  spec: contract.Spec;
-  funcSchema: JSONSchema7;
-}) => {
-  const { network, transaction } = useStore();
-  const { updateSorobanBuildXdr } = transaction;
-  const { isValid } = transaction.build;
-  const { params: txnParams, soroban } = transaction.build;
-  const { operation } = soroban;
-  const {
-    mutate: prepareTx,
-    isPending: isPrepareTxPending,
-    isError: isPrepareTxError,
-    error: prepareTxError,
-    data: prepareTxData,
-    reset: resetPrepareTx,
-  } = useRpcPrepareTx();
-
-  const dereferencedSchema: DereferencedSchemaType = dereferenceSchema(
-    funcSchema,
-    name,
-  );
-
-  useEffect(() => {
-    if (prepareTxData) {
-      updateSorobanBuildXdr(prepareTxData.transactionXdr);
-    } else {
-      updateSorobanBuildXdr("");
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [prepareTxData]);
-
-  const handlePrepareTx = () => {
-    resetPrepareTx();
-
-    const sampleTxnXdr = getTxnToSimulate(
-      value,
-      spec,
-      txnParams,
-      operation,
-      network.passphrase,
-    );
-
-    if (sampleTxnXdr) {
-      prepareTx({
-        rpcUrl: network.rpcUrl,
-        transactionXdr: sampleTxnXdr,
-        networkPassphrase: network.passphrase,
-        headers: getNetworkHeaders(network, "rpc"),
-      });
-    } else {
-      return undefined;
-    }
-  };
-
-  const render = (schema: DereferencedSchemaType): React.ReactElement => {
-    const { name, description } = schema;
-
-    return (
-      <Box gap="md">
-        {renderTitle(name, description)}
-
-        <Box gap="md" key={name}>
-          <JsonSchemaFormRenderer
-            name={name}
-            schema={dereferencedSchema as JSONSchema7}
-            formData={value.args}
-            onChange={onChange}
-          />
-        </Box>
-
-        <Box gap="md" direction="row" wrap="wrap">
-          <Button
-            variant="secondary"
-            disabled={!(isValid.params && isValid.operations)}
-            isLoading={isPrepareTxPending}
-            size="md"
-            onClick={handlePrepareTx}
-            type="button"
-          >
-            Prepare Transaction
-          </Button>
-        </Box>
-
-        {isPrepareTxError && prepareTxError?.result.toString() ? (
-          <ErrorText
-            errorMessage={prepareTxError?.result.toString()}
-            size="sm"
-          />
-        ) : (
-          <></>
-        )}
-      </Box>
-    );
-  };
-
-  return (
-    <Box gap="md">
-      <Card>{render(dereferencedSchema)}</Card>
-    </Box>
-  );
-};
-
-const renderTitle = (name: string, description: string) => (
-  <>
-    <Text size="lg" as="h2">
-      {name}
-    </Text>
-    {description ? <div>{description}</div> : null}
-  </>
-);
-
-const getTxnToSimulate = (
-  value: SorobanInvokeValue,
-  spec: contract.Spec,
-  txnParams: TransactionBuildParams,
-  operation: TxnOperation,
-  networkPassphrase: string,
-) => {
-  try {
-    const scVals = getScValsFromSpec(value.function_name, spec, value);
-    const builtXdr = buildTxWithSorobanData({
-      params: txnParams,
-      sorobanOp: {
-        ...operation,
-        params: {
-          ...operation.params,
-          contract_id: value.contract_id,
-          function_name: value.function_name,
-          args: scVals,
-          resource_fee: BASE_FEE, // bogus resource fee for simulation purpose
-        },
-      },
-      networkPassphrase,
-    });
-
-    return builtXdr.toXDR();
-  } catch (e: any) {
-    console.error("e", e);
-    return undefined;
   }
 };
